requirements.txt
fastapi
uvicorn
joblib
google-cloud-storage
nltk
pandas
scikit-learn
python-dotenv
jinja2
mediacloud-api-legacy